third_party/WebKit/LayoutTests/html/dialog/dialog-canceling-expected.txt

Tests canceling modal dialogs using the Escape key. To test manually, hit Escape once to see the topmost dialog turn green, then once again to close it. Repeat for the remaining dialog.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Top dialog event listener should prevent closing.
PASS topDialog.open is true
PASS bottomDialog.open is true
Top dialog should close.
PASS topDialog.open is false
PASS bottomDialog.open is true
Input should swallow Escape mechanism.
PASS topDialog.open is false
PASS bottomDialog.open is true
Bottom dialog event listener should prevent closing.
PASS topDialog.open is false
PASS bottomDialog.open is true
Bottom dialog should close.
PASS topDialog.open is false
PASS bottomDialog.open is false
Pressing Escape now should do nothing.
PASS topDialog.open is false
PASS bottomDialog.open is false
PASS successfullyParsed is true

TEST COMPLETE
